now store foodborne multiplier when running BuildCostTable.R
This was done in such a way as to preserve the order in which random numbers are drawn, ensuring backwards compatibility. This was done using a global variable storing the foodborne multipliers for the number of incident cases. These multipliers can then be used to mulitply back up to total costs, cases, deaths etc.